refactor(memory): revamp worker launching
