+++ EvoScientist/gateway/runtime.py
@@ -3,10 +3,12 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Literal
 
+from langgraph_sdk import get_client
+from langgraph_sdk.client import LangGraphClient
+
 from .local import LocalGraphGateway, LocalThreadStore
 from .server import (
     DEFAULT_GRAPH_ID,
-    LangGraphClientFactory,
     LangGraphServerGateway,
     LangGraphServerThreadStore,
 )
@@ -29,25 +31,18 @@ def create_runtime_gateways(
     base_url: str | None = None,
     graph_id: str = DEFAULT_GRAPH_ID,
     headers: dict[str, str] | None = None,
-    client_factory: LangGraphClientFactory | None = None,
+    langgraph_client: LangGraphClient | None = None,
 ) -> RuntimeGateways:
     """Create gateway handles for CLI/TUI/serve execution."""
     if backend == "langgraph_server":
-        if base_url is None:
+        if base_url is None and langgraph_client is None:
             raise ValueError("base_url is required for langgraph_server gateways")
-        if client_factory is not None:
-            server_thread_store = LangGraphServerThreadStore(
-                base_url=base_url,
-                graph_id=graph_id,
-                headers=headers,
-                client_factory=client_factory,
-            )
-        else:
-            server_thread_store = LangGraphServerThreadStore(
-                base_url=base_url,
-                graph_id=graph_id,
-                headers=headers,
-            )
+        server_thread_store = LangGraphServerThreadStore(
+            client=langgraph_client
+            if langgraph_client is not None
+            else get_client(url=base_url, headers=headers),
+            graph_id=graph_id,
+        )
 
         return RuntimeGateways(
             thread_store=server_thread_store,
