+++ EvoScientist/memory/agents/__init__.py
@@ -1,13 +1,7 @@
 """Background memory agent implementations."""
 
-from .memory_worker import (
-    alaunch_memory_worker,
-    build_memory_worker_graph,
-    launch_memory_worker,
-)
+from .memory_worker import build_memory_worker_graph
 
 __all__ = [
-    "alaunch_memory_worker",
     "build_memory_worker_graph",
-    "launch_memory_worker",
 ]
